Feature/refator register child specialisst

+++ app/src/main/java/com/example/comusenias/presentation/component/home/CardGame.kt
@@ -60,7 +60,7 @@ fun CardGame(
             verticalAlignment = Alignment.CenterVertically,
             horizontalArrangement = Arrangement.spacedBy(SIZE28.dp)
         ) {
-            ImageWhitBorder(image = subLevel.image.toInt(), borderColor = lineColor)
+            ImageWhitBorder(image = subLevel.image, borderColor = lineColor)
             Row(
                 modifier = Modifier
                     .fillMaxWidth(),
